software/remote/requirements.txt
numpy==2.2.1
opencv-python==4.10.0.84
websockets==14.1
pyserial==3.5
